Refactored the code

+++ crates/engine/src/channel.rs
@@ -77,7 +77,7 @@ impl Channel {
         let sound = &self.sounds[self.sound_index as usize];
         let speed = max(sound.speed, 1);
 
-        if self.tick_count % speed as u32 == 0 {
+        if self.tick_count % speed == 0 {
             if self.note_index >= sound.notes.len() as u32 {
                 self.sound_index += 1;
                 self.note_index = 0;
@@ -115,7 +115,7 @@ impl Channel {
                     tone,
                     (volume as f64 / MAX_VOLUME as f64) * (self.volume as f64 / MAX_VOLUME as f64),
                     effect,
-                    speed as u32,
+                    speed,
                 );
             }
